# Add custom rrule to speed up AD with mapped functions