--- a/Project.toml
+++ b/Project.toml
@@ -12,8 +12,9 @@ ChainRulesCore = "1"
 julia = "1"
 
 [extras]
+ChainRulesTestUtils = "cdddcdb0-9152-4a09-a978-84456f9df70a"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 
 [targets]
-test = ["Documenter", "ForwardDiff"]
+test = ["ChainRulesTestUtils", "Documenter", "ForwardDiff"]