--- a/README.md
+++ b/README.md
@@ -12,7 +12,7 @@ changes for functions that perform a change of variables (like coordinate
 transformations).
 
 `ChangesOfVariables` is a very lightweight package and has no dependencies
-beyond `Base`, `LinearAlgebra` and `Test`.
+beyond `Base`, `LinearAlgebra`, `Test` and `ChainRulesCore`.
 
 ## Documentation