--- a/src/ChangesOfVariables.jl
+++ b/src/ChangesOfVariables.jl
@@ -9,6 +9,7 @@ transformations).
 """
 module ChangesOfVariables
 
+using ChainRulesCore
 using LinearAlgebra
 using Test