--- a/src/with_ladj.jl
+++ b/src/with_ladj.jl
@@ -86,13 +86,18 @@ function _with_ladj_on_mapped(map_or_bc::F, y_with_ladj) where {F<:Union{typeof(
     (y, ladj)
 end
 
-function _with_ladj_on_mapped_pullback(thunked_ΔΩ)
+
+# Need to use a type for this, type inference fails when using a pullback
+# closure over YLT in the rrule, resulting in bad performance:
+struct WithLadjOnMappedPullback{YLT} <: Function end
+function (::WithLadjOnMappedPullback{YLT})(thunked_ΔΩ) where YLT
     ys, ladj = unthunk(thunked_ΔΩ)
-    return NoTangent(), NoTangent(), tuple.(ys, ladj)
+    return NoTangent(), NoTangent(), broadcast((y, l) -> Tangent{YLT}(y, l), ys, ladj)
 end
 
 function ChainRulesCore.rrule(::typeof(_with_ladj_on_mapped), map_or_bc::F, y_with_ladj) where {F<:Union{typeof(map),typeof(broadcast)}}
-    return _with_ladj_on_mapped(map_or_bc, y_with_ladj), _with_ladj_on_mapped_pullback
+    YLT = eltype(y_with_ladj)
+    return _with_ladj_on_mapped(map_or_bc, y_with_ladj), WithLadjOnMappedPullback{YLT}()
 end
 
 function with_logabsdet_jacobian(mapped_f::Base.Fix1{<:Union{typeof(map),typeof(broadcast)}}, X)