AD of `_with_ladj_on_mapped` causes a significant overhead when using AD on `with_logabsdet_jacobian` with mapped/broadcasted functions. This PR adds a custom `ChainRulesCore.rrule` for `_with_ladj_on_mapped` that results in a siginificant speedup. GC overhead is also reduced.

Pro: Very significant speed gain in the benchmark example below. Should be be similar for other lightweight transformations (example uses a `log`-trafo).

Con: The price is adding ChainRulesCore as a dependency

This would make ChangesOfVariables itself less lightweight. On the other hand it's likely that packages that we'd want to depend on ChangesOfVariables will depend on ChainRulesCore already anyway (directly or indirectly).

Benchmark example:

```julia
using ChangesOfVariables, LinearAlgebra, Zygote, BenchmarkTools

function foo(xs)
    ys, ladj = with_logabsdet_jacobian(Base.Fix1(broadcast, log), xs)
    dot(ys, ys) + ladj
end

grad_foo(xs) = Zygote.gradient(foo, xs)[1]

xs = rand(10^3);
grad_foo(xs);

@benchmark grad_foo($xs)
```

Without this PR (no custom rrule):

```julia
julia> @benchmark grad_foo($xs)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  30.507 μs …   2.065 ms  ┊ GC (min … max):  0.00% … 95.86%
 Time  (median):     46.969 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   58.705 μs ± 122.055 μs  ┊ GC (mean ± σ):  15.00% ±  7.01%

         ▁▆█▇▄▃▃▂▁▁                                            ▁
  ▅▄▆██▆▇█████████████▇▆▅▅▄▅▄▆▆▆▆▆▅▅▆▆▆▇▆▇▆▅▆▆▆▆▆▆▅▄▅▅▅▃▅▃▄▄▄▃ █
  30.5 μs       Histogram: log(frequency) by time       130 μs <

 Memory estimate: 248.20 KiB, allocs estimate: 99.
```

With this PR (custom rrule for `_with_ladj_on_mapped`):

```julia
julia> @benchmark grad_foo($xs)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  19.175 μs …  2.898 ms  ┊ GC (min … max):  0.00% … 93.90%
 Time  (median):     27.598 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   34.333 μs ± 81.148 μs  ┊ GC (mean ± σ):  12.31% ±  5.25%

         ▁▄▇█▇▅▄▃▃▂▁▁▁ ▁▁▂▃▃▃▂▂▁                              ▂
  ▄▆▅▃▃▅▇█████████████████████████▇▇▇▇▇▆▆▅▆▅▅▆▄▅▄▄▅▄▅▄▃▄▄▄▅▄▄ █
  19.2 μs      Histogram: log(frequency) by time      64.4 μs <

 Memory estimate: 136.44 KiB, allocs estimate: 50.
```