@oschulz 2021-11-08
> Is the Zygote gradient with map faster or is it a broadcast-specific issue?

It's slow with `sum(map(last, y_with_ladj))`, `sum(broadcast(last, y_with_ladj))` and, `sum(last, y_with_ladj)`. :-(