@oschulz 2021-12-09
The package that currently depend on ChangesOfVariables already depend on ChainRulesCore directly or indirectly, so adding it as a dependency here doesn't affect their load time (just tested for LogExpFunctions and TransformVariables, just to make sure).