@oschulz 2021-12-10
> * Are the benchmarks in the PR still correct?

Looks like it:

Current master:

```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  31.854 μs …   2.385 ms  ┊ GC (min … max):  0.00% … 94.80%
 Time  (median):     43.974 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   59.229 μs ± 146.713 μs  ┊ GC (mean ± σ):  18.31% ±  7.22%
```

This PR:

```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  18.304 μs …   2.671 ms  ┊ GC (min … max):  0.00% … 97.44%
 Time  (median):     25.303 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   34.384 μs ± 105.128 μs  ┊ GC (mean ± σ):  16.69% ±  5.42%
```

> * Would it be useful to add a `frule` as well?

I don't think it would bring a speedup (at least not with ForwardDiff, which doesn't support `frule` anyway, we'd need a Dual rule and depend on it). Maybe hold off on that until we can benchmark it (when a major AD framework supports `frules`s)?